@@ -1,47 +1,74 @@
 {
   "name": "vue-validator",
   "description": "Validator component for Vue.js",
-  "version": "1.4.4",
-  "author": "kazuya kawaguchi",
+  "version": "2.0.0-alpha",
+  "author": {
+    "name": "kazuya kawaguchi",
+    "email": "kawakazu80@gmail.com"
+  },
   "bugs": {
     "url": "https://github.com/vuejs/vue-validator/issues"
   },
   "devDependencies": {
-    "commander": "^2.8.1",
-    "eslint": "^0.24.0",
-    "expect.js": "^0.3.1",
-    "karma": "^0.12.31",
-    "karma-chrome-launcher": "^0.1.7",
-    "karma-commonjs": "0.0.12",
-    "karma-coverage": "^0.2.7",
-    "karma-coveralls": "^0.1.4",
-    "karma-expect": "^1.1.0",
-    "karma-firefox-launcher": "^0.1.4",
-    "karma-mocha": "^0.1.9",
-    "karma-phantomjs-launcher": "^0.1.4",
+    "babel": "^5.8.23",
+    "babel-core": "^5.8.23",
+    "babel-loader": "^5.3.2",
+    "babel-plugin-espower": "^1.0.0",
+    "commander": "^2.9.0",
+    "eslint": "^1.3.1",
+    "espower-babel": "^3.3.0",
+    "istanbul-instrumenter-loader": "^0.1.3",
+    "json-loader": "^0.5.2",
+    "karma": "^0.13.9",
+    "karma-chrome-launcher": "^0.2.0",
+    "karma-coverage": "^0.5.2",
+    "karma-coveralls": "^1.1.2",
+    "karma-firefox-launcher": "^0.1.6",
+    "karma-mocha": "^0.2.0",
+    "karma-mocha-reporter": "^1.1.1",
+    "karma-phantomjs-launcher": "^0.2.1",
     "karma-safari-launcher": "^0.1.1",
-    "karma-sauce-launcher": "^0.2.10",
-    "karma-sinon": "^1.0.4",
-    "mocha": "^1.18.2",
-    "nightmare": "^1.6.5",
-    "semi": "^3.1.1",
-    "sinon": "^1.15.4",
-    "vue": "^0.12.16",
-    "webpack": "^1.4.13"
+    "karma-sauce-launcher": "^0.2.14",
+    "karma-sourcemap-loader": "^0.3.5",
+    "karma-webpack": "^1.7.0",
+    "mocha": "^2.3.1",
+    "nightmare": "^2.0.5",
+    "phantomjs": "^1.9.18",
+    "power-assert": "^1.0.0",
+    "vue": "^1.0.8",
+    "webpack": "^1.12.1",
+    "webpack-dev-server": "^1.11.0"
   },
   "homepage": "https://github.com/vuejs/vue-validator",
   "keywords": [
     "plugin",
     "validation",
-    "vue"
+    "vue",
+    "vue.js"
   ],
   "license": "MIT",
-  "main": "index.js",
+  "main": "dist/vue-validator.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/vuejs/vue-validator.git"
   },
   "scripts": {
-    "test": "make ci"
+    "browser": "VUE_VALIDATOR_TYPE=browser karma start config/karma.conf.js",
+    "build": "babel src --out-dir lib --source-maps-inline",
+    "bundle": "webpack --config config/webpack.conf.js && webpack --config config/webpack.min.conf.js",
+    "ci": "npm run coverage && npm run coveralls && npm run sauce && npm run e2e",
+    "clean": "rm -rf lib && rm -rf coverage && rm -rf dist",
+    "coolkids": "VUE_VALIDATOR_TYPE=sauce SAUCE=batch1 karma start config/karma.conf.js",
+    "coverage": "VUE_VALIDATOR_TYPE=coverage karma start config/karma.conf.js",
+    "coveralls": "VUE_VALIDATOR_TYPE=coveralls karma start config/karma.conf.js",
+    "dev": "webpack-dev-server --inline --quite --hot --config config/webpack.dev.conf.js --content-base ./",
+    "dist": "npm run build && npm run bundle",
+    "e2e": "webpack-dev-server --quiet -h --config config/webpack.e2e.conf.js --content-base test/e2e & mocha -t 20000 --compilers js:espower-babel/guess test/e2e/test.js && kill $! || (kill $! && exit 1)",
+    "ie": "VUE_VALIDATOR_TYPE=sauce SAUCE=batch2 karma start config/karma.conf.js",
+    "lint": "eslint src test/specs config",
+    "mobile": "VUE_VALIDATOR_TYPE=sauce SAUCE=batch3 karma start config/karma.conf.js",
+    "sauce": "npm run coolkids && npm run ie && npm run mobile",
+    "test": "karma start config/karma.conf.js",
+    "watch": "babel src --out-dir lib --watch --source-maps-inline"
   }
 }
